+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 0.9.5
-
+- New: Move password to different folder
+- New: Delete folder
 
 ## 0.9.4
 - Add yes/no dialog for logout